refac: modularize maps.py
