+++ pypsa/plot/maps/__init__.py
@@ -1,0 +1,16 @@
+"""
+Map plots for PyPSA.
+
+This module contains all plotting functions which can be used on top of the
+functions of :mod:`pypsa.plot.map`, :mod:`pypsa.plot.explore`, :mod:`pypsa.plot.iplot`.
+"""
+
+from pypsa.plot.maps.interactive import explore, iplot
+from pypsa.plot.maps.static import MapPlotter, plot
+
+__all__ = [
+    "MapPlotter",
+    "plot",
+    "iplot",
+    "explore",
+]
