+++ test/test_plot_maps.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from pypsa.plot.maps import (
+from pypsa.plot.maps.static import (
     add_legend_circles,
     add_legend_lines,
     add_legend_patches,
